app/requirements.txt
# Core dependencies
flask>=3.0.0
gunicorn>=21.0.0

# Data processing
pandas>=2.0.0

# Web scraping
playwright>=1.40.0

# Database
supabase>=2.3.0

# Excel file processing
openpyxl>=3.1.0

# HTTP requests (for testing)
requests>=2.31.0

# Date manipulation
python-dateutil>=2.8.0 